Move test to test_fervo_project_cape_4.py
